Addition of Jacobi and Legendre polynomials and Vandermonde matrix
Legendre polynomials, evaluated as Jacobi polynomial with alpha = 0 and beta = 0, constitute the optimal modal basis for the elements.

The Vandermonde matrix acts as a linear operator that maps a vector  expressed in the basis of Legendre polynomials in to the nodal basis of Lagrange polynomials, which are used by default in this solver.

The main point of implementing Vandermonde matrix is because it acts a pivot to build the exact mass matrix (no Legende-Gauss-Lobatto quadrature) and this feature will be used to quickly build the 2D mass matrix, which will also use Lagrange polynomials as the nodal basis